Estimate cut-off values for different transmission types

Estimate cut-off values for different transmission types

get_quantiles_multi(
  d_type,
  distrib,
  obs,
  rr,
  n,
  params,
  q,
  assort_mix,
  prop_type = "all"
)

Arguments

d_type

The distance type - one of either temporal or spatial

distrib

The distribution to be used for estimation. For d_type = temporal, can be lognormal or gamma. For d_type = spatial, can be Rayleigh or gamma

obs

Number of observations in each group. Should be entered as a vector of length 2 (group 1 obs, group 2 obs)

rr

Reporting probabilities for each group. Should be entered as a vector of length 2 (group 1 reporting probability, group 2 reporting probability)

n

Number of individuals to include in simulation. Recommend >1,000,000

params

The summary statistics for the distance type specified in d_type. Entered as a vector of length 6. For d_type = spatial it will be (mean for g1-g1 transmission, std for g1_g1 transmission, mean for mixed transmission, std for mixed transmission, mean for g2-g2 transmission, std for g2-g2 transmission)

q

The centile to use for the cut-off e.g 0.95 for 95th centile.

assort_mix

The level of assortative mixing where 1 represents random mixing.

prop_type

The transmission proportions to extract from the simulation. Options are "all" (default) in which the proportions from the whole simulation are extracted and "centile" in which only the transmissions below the centile cut off value are used to extract the proportions of each transmission type

Value

A dataframe containing the cut-off values for each type of transmission and the proportion of each type of transmission that were "observed" in the simulation

Examples

g1_obs <- 300
g2_obs <- 200
g1_rr <- 0.75
g2_rr <- 0.50
n = 10000000
q <- 0.95
si_mean <- 27.8
si_sd <- 36.8
params_temporal <- c(si_mean, si_sd, si_mean, si_sd, si_mean, si_sd)
assort_mix <- 1

get_quantiles_multi(d_type = "temporal", distrib = "lognormal",
                    obs = c(g1_obs, g2_obs), rr = c(g1_rr, g2_rr),
                    params = params_temporal,
                    n = n, q = q, assort_mix = assort_mix, prop_type = "all" )
#> Error in get_quantiles_multi(d_type = "temporal", distrib = "lognormal",     obs = c(g1_obs, g2_obs), rr = c(g1_rr, g2_rr), params = params_temporal,     n = n, q = q, assort_mix = assort_mix, prop_type = "all"): unused argument (prop_type = "all")